tiny-tracks-app/server/requirements.txt
asgiref==3.8.1
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
defusedxml==0.8.0rc2
distlib==0.3.8
Django==5.0.4
django-cors-headers==4.3.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
exceptiongroup==1.2.1
filelock==3.13.1
Flask==3.0.0
Flask-Cors==4.0.0
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.26.2
oauthlib==3.2.2
opencv-python==4.8.1.78
packaging==24.0
pipenv==2023.11.17
platformdirs==4.1.0
pluggy==1.5.0
pycparser==2.22
PyJWT==2.8.0
PySimpleGUI==4.60.5
pytest==8.1.1
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==2.0.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
tabulate==0.9.0
tomli==2.0.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.1.0
virtualenv==20.25.0
Werkzeug==3.0.1
